requirements.txt
matplotlib==3.10.6
numpy==2.3.2
opencv_python==4.12.0.88
Pillow==11.3.0
Requests==2.32.5
scikit_learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
ultralytics==8.3.82


streamlit
pandas
plotly
supabase
